Extract ElasticSearchHelper from consumer search by postcode
