+++ spec/features/consumer_searches_by_postcode_only_spec.rb
@@ -86,22 +86,4 @@ RSpec.feature 'Consumer searches by postcode only' do
   def then_i_am_told_the_postcode_is_incorrect
     expect(landing_page.in_person).to be_invalid_postcode
   end
-
-  def with_fresh_index!
-    `curl -XDELETE -sS http://127.0.0.1:9200/rad_test`
-    `curl -XPOST -sS http://127.0.0.1:9200/rad_test -d @elastic_search_mapping.json`
-    yield
-    `curl -XPOST -sS http://127.0.0.1:9200/rad_test/_refresh`
-  end
-
-  def with_elastic_search!
-    VCR.turned_off do
-      begin
-        WebMock.allow_net_connect!
-        yield
-      ensure
-        WebMock.disable_net_connect!
-      end
-    end
-  end
 end
