+++ spec/spec_helper.rb
@@ -15,4 +15,5 @@ RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
 
   config.include FactoryGirl::Syntax::Methods
+  config.include ElasticSearchHelper, type: :feature
 end
